[prep CDF-25832] 🚧Remove private cognite-sdk depedency (#2027)
# Description

We need to support storing function code in CogniteFile on
DataModelingOnly CDF projects. This is a preparation PR were I remove
the dependency on the private method in the `cognite-sdk` that zips and
uploads the file. This is because depending on a private method is not
good, and it is needed as that private method has asset-centric file
hardcoded.


## Changelog

- [ ] Patch
- [x] Skip