[CDF-25835] ☔ Remove update_items from StorageIO  (#2024)
# Description

In Toolkit we have a `StorageIO` class as an interface for working with
data. Currently it has two methods (`upload_items` and
`upload_items_force`) for uploading data. This PR does the following:

1. Deletes the abstract method `upload_items` from the StorageIO
interface along wiht all implementations of it.
2. Updates tests.
3. Implements `upload_items_force` for files as `/files` is s special
case that only supports creating one file at a time.
4. Rename `upload_items_force` to `upload_items`.

## Changelog

- [ ] Patch
- [x] Skip