[CDF-25868] 🫑Lookup Events/FileMetadata internal ↔ external IDs (#2039)
# Description

This PR adds an internal method for looking up internal IDs and external
IDs from the opposite for events and filemedata.

This is in preparation for supporting download and upload of Canvases.
Canvas references events, file metadata, assets, and timeseries with
internal IDs. When we download them we need to replace them with
external IDs and the reverse in upload. This is such that the user can
download a Canvas in one CDF project and upload it into another.

## Changelog

- [ ] Patch
- [x] Skip