[CDF-25862] 🎨 CanvasIO (#2036)
# Description

**Context** The StorageIO is a generalized interface for interactive
with data in Toolkit. It is used in our download/upload/purge/migrate
commands.

This PR prepares for supporting download/upload of Canvas by doing the
following steps:
1. Make the data classes `IndustrialCanvas` a CogniteResource, this is
required to be used in the `StorageIO` intreface.
2. Create an empty shell for `CanvasIO` with `NotImplemented` errors for
all methods.
3. Move out the fixture for a full Canvas used in the migration testing
such that it can be reused in the upcoming download and upload command
testing.

## Changelog

- [ ] Patch
- [x] Skip