[CDF-25055] ⏫ Cognite extractor mapping (#2040)
# Description


**Context**: This is part of the `alpha` migration tooling in Toolkit,
for overview [see
here](https://cognitedata.atlassian.net/wiki/spaces/~170890702/pages/5455773697/Migration+Tooling+in+Toolkit).
This is step 3 of that process.

This PR adds a default mapping for cognite extractor timeseries/files.


## Changelog

- [x] Patch
- [ ] Skip

## cdf

### Added

- [alpha] (Migration) Adds mapping for CogniteExtractorTimeSeries and
CogniteExtractorFiles. This is deployed when you run `cdf migrate
prepare` and will be available to use in the upcoming improvement of
`cdf migrate timeseries/files`.

## templates

No changes.