[CDF-25802] 😃 Infield v1 YAML (#2051)
# Description

We are in the progress of adding pydantic classes to match the YAML
format Toolkit expects for configurations. The goal is to give better
error message to the user on syntax errors.

This PR introduces the `InfieldV1YAML` to validate infield
configurations. Note this is not yet exposed, so no change to the end
user. Also note that the InfieldV1 resource is an alpha feature.

## Changelog

- [ ] Patch
- [x] Skip