😁 Use YAML classes (#2061)
# Description

These classes are already implemented and well tested. Setting the
yaml_cls in the CRUD interface will start using this in the build
command to validate the YAML syntax of the user. The motivation for
doing this in a separate PR is to avoid merge conflicts.

This is the same type of PR as for example #1944, #1994

**NOTE** This is the last PR of this type. We are done with creating
YAML classes, so I have added a test to ensure that all ResourceCrud
have the `yaml_cls` set.

## Details:
The testing of the YAML resource classes we use to validate the syntax
of the user configurations:
https://github.com/cognitedata/toolkit/tree/main/tests/test_unit/test_cdf_tk/test_resource_classes

Then, by setting yaml_cls in the resource CRUDit will be used in the cdf
build command to give the user a warning about whether they follow the
syntax here:
https://github.com/cognitedata/toolkit/blob/main/cognite_toolkit/_cdf_tk/commands/build_cmd.py#L603


## Changelog

- [x] Patch
- [ ] Skip

## cdf

### Improved

- Warning on YAML syntax errors when running cdf build for resources
types (alpha features) agents and infield configurations.

## templates

No changes.