[CDF-25913]👷Rework selectors (#2058)
# Description

**Context** We are working towards adding a new data plugin in Toolkit.
This plugin is to support download, upload and purging of data. The
interface we use to implement the logic is called StorageIO. In
StorageIO, we have a concept called selector which is used to select
which data to download/upload/purge.

This PR reworks the selector classes from dataclass to pydantic. This is
to take advantage of the serialization/deserialization capability of
pydantic such that the selector can be stored in a YAML file next to the
data that has been downloaded.

## Changelog

- [ ] Patch
- [x] Skip