[CDF-25913] 🏛 Implement more selectors (#2060)
# Description

**Context** We are working towards adding a new data plugin in Toolkit.
This plugin is to support download, upload and purging of data. The
interface we use to implement the logic is called StorageIO. In
StorageIO, we have a concept called selector which is used to select
which data to download/upload/purge.

This PR builds on #2058, it implements the remaining selectors in their
pydantic version.

## Changelog

- [ ] Patch
- [x] Skip