[CDF-25914] 🏗 Use new selectors (#2059)
# Description

**Context** We are working towards adding a new data plugin in Toolkit.
This plugin is to support download, upload and purging of data. The
interface we use to implement the logic is called StorageIO. In
StorageIO, we have a concept called selector which is used to select
which data to download/upload/purge.

The PR #2058 reimplemented selectors into pydantic classes, this PR uses
replaces the previous data class based selectors with then new pydantic
based selectors in the StorageIO interface. Note there is no change in
behavior for the user.

## Changelog

- [ ] Patch
- [x] Skip