[CDF-25931] Bug fix for AppConfigAcl in auth verify (#2077)
# Description

When running `cdf auth verify` after enabling `search-config` alpha
feature, users are getting error `TypeError: unhashable type: 'list'`.
This PR intends to fix that error which is due to list in
`AppConfigScope(apps=['SEARCH'])`. In this PR we are converting the
scope object to set of tuples with scope data and a reverse mapping is
saved for look-up to avoid this error when running `cdf auth verify`.

## Changelog

- [x] Patch
- [ ] Skip

## cdf

### Fixed

- fixed `TypeError: unhashable type: 'list'` error on running `cdf auth
verify`

## templates

No changes.