[prepare CDF-25905, CDF-25906]🐛 Selectors are hashable (#2082)
# Description

**Context** We are working towards adding a new data plugin in Toolkit.
This plugin is to support download, upload and purging of data. The
interface we use to implement the logic is called StorageIO. In
StorageIO, we have a concept called selector which is used to select
which data to download/upload/purge.

For the upload command selectors must be hashable so they can be used as
a key in a mapping. This is the case for all but one selector. This PR
extends the testing to check for hashable and fixes the one case.

## Changelog

- [ ] Patch
- [x] Skip