[prep CDF-25905, CDF-25906] 😄 Instance download upload logic (#2083)
# Description

**Context** We are working towards adding a new data plugin in Toolkit.
This plugin is to support download, upload and purging of data.

This PR adds the logic for downloading and uploading instances. 

I introduce a new selector `InstanceSpaceSelector`, this is because when
downloading intances I think it is natural to select an instance space,
and then which views to download properties from.

## Changelog

- [ ] Patch
- [x] Skip