[prep CDF-25905, CDF-25906] 📇 InstanceIO  (#2085)
# Description

**Context** We are working towards adding a new data plugin in Toolkit.
This plugin is to support download, upload and purging of data.

This PR changes the InstanceIO to support configurations. This is used
in the upcoming download and upload command. When you download
instances, then spaces, views and containers are downloaded
automatically. These are then optimally deployed when the instances are
uploaded.

## Changelog

- [ ] Patch
- [x] Skip