🧹StorageIO remove folder name (#2090)
# Description

The StorageIO is an interface for working with data
(download/upload/purge) in Toolkit.

This PR removes the `FOLDRE_NAME` from the interface, as this is no
used. In the recent redesign #2069 , data is instead downloaded to a
folder describing the data for example dataset external id, instance
space, or raw databse.


## Changelog

- [ ] Patch
- [x] Skip