[CDF-25979] 😆 Migration configuration creation  (#2099)
# Description

**Context**: This is part of the `alpha` migration tooling in Toolkit,
for overview [see
here](https://cognitedata.atlassian.net/wiki/spaces/~170890702/pages/5455773697/Migration+Tooling+in+Toolkit).
This is a new step between 2 and 3.

Before we can migrate assets/event/files/timeseries/sequences, we need
to have a destination target space as well as CogniteSourceSystems
created.

This PR introduces a new method in the MigrationCommand with a creator
interface to create InstanceSpaces from datasets and CogniteSourceSystem
from asset/event/file `source` property. Instance space and cognite
source system are treated as configuration and not data, thus I use the
deploy command to create them and output the configurations such that
the user can add it to their governance.

## Changelog

- [ ] Patch
- [x] Skip