[prep CDF-25991] 😄Rework StorageIO (#2110)
# Description

**Context** In Toolkit, we have a data plugin for
downloading/uploading/purging data. Internally we use the interface
`StorageIO` for doing this operation.

This PR splits the `StorageIO` into `StorageIO` and
`UploadableStorageIO`, in other words, it allows us to create
implementation of this interface that does not support upload. This such
that we can introduce `HierarchyIO` which is an implementation that
downloads assets/events/files metada/timeseries. In addition, I have
moved the `KIND` to the selectors (what is used to select the data to
download) to support the `HierarchyIO` case.

## Changelog

- [ ] Patch
- [x] Skip