[CDF-25485] 🚧Remove stop-gap solution for Function Create endpoint queuing (#2115)
# Description

CogniteFunctions response with a 429 and `retry-after` in the header
that is not supported by the Cognite-SDK. Thus, we have a special
implementation in Toolkit to handle this type of status.

Before doing the fix for the delete endpoint, I refactored the stop-gap
solution for the create that uses custom logic to handle the response,
and instead use the `HTTPClient` implemented in Toolikt to do it. In
addition, I updated the `HTTPClient` to specially handle the 429
responses.

## Changelog

- [ ] Patch
- [x] Skip