[prep CDF-25991] 😆 Rework http client (#2119)
# Description

This PR reworks the `HTTPClient` due to the following

* The HTTPMessage type `ItemsRequest` assumes that you can use the same
function for obtaining ID from an item from a request and a response.
This is not true. For the upload functionality of `cdf data upload` the
request ID will be a line number in a file.
* The `ItemsRequest` assumes shape of the request. The new version does
not. This enables us to simply delete classes such as
`UnknownResponseItem`.
* Removed all serialization from the `HTTPClient`. This makes the client
a single responsibility class for making requests and returning
responses. This means if we want to use pydantic for serialization in
the future we can.
* For the `RequestMessage` now have two methods for creating responses
instead of one, `create_success_response` and `create_failure_response`.
This allows us to not duplicate the logic for checking response code
twice, once in the `HTTPClient` and one in the create_response method.
* Instead of individual ResponseItem classes we now have containers.
This is that if we do a request with 1000 items and get a 200 response,
we can assume all items were successfully set. Similar if all fail or
the request fails.
 
## Changelog

- [ ] Patch
- [x] Skip