[prep CDF-25991] 😆 Rework StorageIO part 2 try 2 (#2120)
# Description

This PR introduces the following changes to the StorageIO interfaces:

1. The id is hardcode to `string` as it is only used for logging. This
simplifies the `as_id` function to -
https://github.com/cognitedata/toolkit/pull/2120/files#diff-2ab3a7cd764ca1cf9f9f669767652ac2faa76240c3cef2ea60c9004626cbdabeR60
2. The `stream_data` method now returns a `Page` in preparation for
storing the cursor.
https://github.com/cognitedata/toolkit/pull/2120/files#diff-2ab3a7cd764ca1cf9f9f669767652ac2faa76240c3cef2ea60c9004626cbdabeR94
3. The `StorageIO` is now a generic of only two types, the selector used
to chose data and an individual data item. Then, the
`data_to_json_chunk` has to be abstract as we do not know that we have
an `.as_write` method anymore:
https://github.com/cognitedata/toolkit/pull/2120/files#diff-2ab3a7cd764ca1cf9f9f669767652ac2faa76240c3cef2ea60c9004626cbdabeR119
4. The `upload_items` now have an `UploadItem` dataclass that wraps the
item wiht an ID. This allows us to track the source of the item from for
example a line in a file.
https://github.com/cognitedata/toolkit/pull/2120/files#diff-2ab3a7cd764ca1cf9f9f669767652ac2faa76240c3cef2ea60c9004626cbdabeR151
and
https://github.com/cognitedata/toolkit/pull/2120/files#diff-2ab3a7cd764ca1cf9f9f669767652ac2faa76240c3cef2ea60c9004626cbdabeR181


The rest this PR is implementing the consequence of these interface
changes.


## Changelog

- [ ] Patch
- [x] Skip