[CDF-26067] 🧱 Select consumption view (#2123)
# Description

**Context**: This is part of the `alpha` migration tooling in Toolkit,
for overview [see
here](https://cognitedata.atlassian.net/wiki/spaces/~170890702/pages/5455773697/Migration+Tooling+in+Toolkit).
This is step 3 of that process.

This PR extends the interactive selection of views to take mapped
containers. This way I can for example pass in `ContainerId("cdf_cdm" ,
"CogniteAsset")` to only select views that extents CogniteAsset. I will
use this to enable the user to select which view to prefer for
consumption.


## Changelog

- [ ] Patch
- [x] Skip