[CDF-25293] 🏭 Migrate Assets (#2095)
# Description


Stacked on #2127 and #2126 

**Context**: This is part of the `alpha` migration tooling in Toolkit,
for overview [see
here](https://cognitedata.atlassian.net/wiki/spaces/~170890702/pages/5455773697/Migration+Tooling+in+Toolkit).
This is step 3 of that process.

This is the first of the 'v2' migration commands. This uses the new
generic interface for migrating assets.

Note I did end-to-end testing of the command, and by doing that I found
a few issues noted in point 3-7 below. I chose to include these
fixes/improvements/features in this PR and not separate them out, as I
think they should be considered in context of the entire command.

Changes in this PR
1. Adding a CLI interface for running `cdf migrate assets`.
[Ref](https://github.com/cognitedata/toolkit/pull/2095/files#diff-f1d451d71388110ca4f82b7c99aee1647984344110f942e08425acd61b371aaaR230)
2. Introducing a new selector `MigrateDataSetSelector` (with
implementation) to find which assets to migrate and how to migrate them.
[Ref](https://github.com/cognitedata/toolkit/pull/2095/files#diff-6a405b5fb033aa8474a3bd402266260bff81d8c0c17e96275ec1fab202aa073fR80).
Used in the interactive selection of assets to migrate as well as an
alternative option to CSV file with all the IDs.
3. Fixing type hints for resource type:
[Ref](https://github.com/cognitedata/toolkit/pull/2095/files#diff-7c3332d7619c62401ad144b404cb754c52e09a60be2b3ec7af4d6d30f279e463R31).
Robustifying.
4. I did some end-to-end tests, so found a few issues with the `cdf
migrate source-systems` command. It was not plural like all other
commands, and should give the option to select an empty instance space:
[Ref](https://github.com/cognitedata/toolkit/pull/2095/files#diff-f1d451d71388110ca4f82b7c99aee1647984344110f942e08425acd61b371aaaR154)
5. During debugging added traceback to the worker producer. Without it,
it gets very hard to find the source of errors:
[Ref](https://github.com/cognitedata/toolkit/pull/2095/files#diff-)c48189d7a95f258d8b5995127e72f5bddbef7ab7bb32e4e0c88a1e247cf1a7efR195
6. Added missing mapping from `parentId` to `partent` in the default
asset mapping.
[Ref](https://github.com/cognitedata/toolkit/pull/2095/files#diff-0cce24ab97b9ee3cb0d8cff2e8d0da4cf15a1ccfb948df2399508892943f966fR48)
7. Added lookup of all assets to support adding `parent`.
[Ref](https://github.com/cognitedata/toolkit/pull/2095/files#diff-724cb4682263702cde512160242469a76c136f464535c695b7ad1eacc07cd988R86)

<img width="1437" height="466" alt="image"
src="https://github.com/user-attachments/assets/1117c4b2-782e-4a01-9377-0b4fd51a08fa"
/>

<img width="1869" height="278" alt="image"
src="https://github.com/user-attachments/assets/d7d718ac-fadd-487e-a9f5-697550406fe6"
/>

## Changelog

- [x] Patch
- [ ] Skip

## cdf

### Added

- [alpha] Adds new command for migrating assets `cdf migrate assets`. 

### Changed

- [alpha] The migration command for source system is renamed from `cdf
migrate source-system` to `cdf migrate source-systems` to be consistent
with the other migration commands.

## templates

No changes.