[CDF-26105] 😉Prepare phase2 Part 2 (#2140)
# Description

Renamed `AssetCentricMigrationIOAdapter` to `AssetCentricMigrationIO`
and simplified the implementation by basing it on the `HierarchyIO` .
The new implementation made the `TimeSeriesIOAdapter` and
`FilesIOAdapter` obsolete.

Note that the new `AssetCentricMigrationIOAdapter` will make
implementing a `cdf migrate hierarchy` much simpler as that now supports
Events, TimeSeries, Assets, and Files only using different selectors to
get the different types.

## Changelog

- [ ] Patch
- [x] Skip