[CDF-25834] 🏛 Rework purge dataset Part 3 (#2147)
# Description

This is part 1 of reimplementing the cdf purge dataset command. 

This PR implements the v2 of purge dataset.

The motivation for this reimplementation is as follows:
* We have new components such as HTTPClient and WorkerProducer that will
be very useful for running the purge code.
* The current implementation lacks proper testing, a reimplementation
will ensure it is well tested.

<img width="568" height="434" alt="image"
src="https://github.com/user-attachments/assets/cb22635b-258c-4105-ba84-bcfff040fe1d"
/>


## Changelog

- [ ] Patch
- [x] Skip