[prep CDF-25023] 😄 Refactor migration code prepare for annotations. (#2161)
# Description

My original intention was to have a `BaseMigrationCommand` and subclass
it for all individual commands. Now, I have one command and instead use
different implementations of `StorageIO` for the different
implementations.

This PR removes the `BaseMigrationCommand`. The only subclass it has is
the `CanvasMigrationCommand`, so I just moved them methods into that
one. Once we support download/upload Canvas using an StorageIO
interface, I will remove the CanvasMigrationCommand as well.

## Changelog

- [ ] Patch
- [x] Skip