[prep CDF-25023]  🙃 Type migration conversion (#2162)
# Description

This is preparation for supporting `cdf migrate annoations`. 

In this PR, I specify the type cognite resource to be asset, file,
event, sequence or timeseries in the migration code. Before it was just
a generic CogniteResoruce that could be anything. I do this to have type
safety when I introduce annotations as one of the supported types.

## Changelog

- [ ] Patch
- [x] Skip