[prep CDF-26155] 🏚 Moving out schema sniffing logic from CSVReader (#2167)
# Description

This is the preparation for adding SQLLite reader/writer. This also
needs the ability to sniff a schema as we can only store null, integer,
real, text and blob in the SQLLite database. The file readers/writers
are expected to handle date/datetime/json as well as the basic types.

## Changelog

- [ ] Patch
- [x] Skip