[CDF-26172, CDF-26170]  😃Infield CDM V1 CRUD (#2181)
# Description

Adds support for the Infield CDM V1 CRUD.

We are moving away from depending on internals of the `cognite-sdk`, and
this is the first resource used in a CRUD implementation that is not
using the `CogniteResource` as defined by the `cognite-sdk`. To support
the new resource classes, I introduced [these
protocols](https://github.com/cognitedata/toolkit/pull/2181/files#diff-beedbbdc48da38bb10f4d3a05c2db6dd2b680d677ee725dfc7b0268af6b287bbR14)
. This is ducktyping of the resources used by the CRUD classes such that
we not have to subclass `CogniteResource`, `CogniteResourceList`, and so
on. Note if we decide to remove the `cognit-sdk` as a dependency these
protocols can be replaced by the pydantic classe ResoruceRequest and
ResourceResponse.


## Changelog

- [x] Patch
- [ ] Skip

## cdf

### Added

- [alpha] Support for Infield Location Config in the `cdf build/deploy`
methods.

## templates

No changes.