[CDF-25328] Upload: Add filtering for read-only properties in data modeling (#2166)
# Description

This PR adds automatic filtering of read-only properties when uploading
data model instances. A new method `_filter_readonly_properties`,
detects and removes read-only properties from instance data before
upload, preventing API errors that occur when attempting to write to
properties like `assetHierarchy_path`. This is solved by using a
constant with a list of the relevant properties, since the API currently
does not expose whether a property is read-only or not.

## Changelog

- [x] Patch
- [ ] Skip

## cdf

### Improved

- [alpha] When uploading instances using the `cdf data upload` command,
Toolkit now skips read-only properties as these cannot be set through
the API. This means, for example, CogniteAsset and CogniteFile instances
can now be uploaded and will not fail anymore.

## templates

No changes.

---------

Co-authored-by: Magnus Schjølberg <magnus.schjolberg@cognitedata.com>