[CDF-26307] 👷 Mapping in chunks (#2208)
# Description

**Context**: This is part of the `alpha` migration tooling in Toolkit,
for overview [see
here](https://cognitedata.atlassian.net/wiki/spaces/~170890702/pages/5455773697/Migration+Tooling+in+Toolkit).
This is step 4 of that process.


This PR refactors the lookup of direct relations in data modeling from
internal/externalId asset-centric.

1. Refactor the mapping in the migration to be done in chunks. This is
to be able to do a bulk look-up of references resources.
[Ref](https://github.com/cognitedata/toolkit/pull/2208/files#diff-724cb4682263702cde512160242469a76c136f464535c695b7ad1eacc07cd988R87).
2. Rewrite the direct relation cache such that it can do bulk updates,
[ref](https://github.com/cognitedata/toolkit/pull/2208/files#diff-dea81a2f4648a92b7eb9adfc619cf0fd1e7bf913e7c743480a26b21db4c0886cR65)
3. I also extended the direct relation cache with asset external ID,
file ID and external ID. This is to support annotation migration.


## Changelog

- [ ] Patch
- [x] Skip