[CDF-24983] 😝 Converting Charts from Asset-Centric to DM (#2219)
# Description

**Context**: This is part of the `alpha` migration tooling in Toolkit,
for overview [see
here](https://cognitedata.atlassian.net/wiki/spaces/~170890702/pages/5455773697/Migration+Tooling+in+Toolkit).
This is step 4 of that process.


This PR adds the conversion from an asset-centric Chart to a DM Chart.

The conversion logic is in the method `_map_single_item`
[here](https://github.com/cognitedata/toolkit/pull/2219/files#diff-724cb4682263702cde512160242469a76c136f464535c695b7ad1eacc07cd988R163)

Found a bug in the serialization of charts
[here](https://github.com/cognitedata/toolkit/pull/2219/files#diff-12bfa38dcaf3415d3d0c139f33a9561617b466aa972a5444526f42fd790e1218R138),
which is the reason for the release.

The migration is not yet released, that will the next PR which will add
a CLI interface which will use the logic implemented in this PR.

## Changelog

- [x] Patch
- [ ] Skip

## cdf

### Fixed

- [alpha] The download chart command, `cdf data download charts`, no
longer fails to serialize

## templates

No changes.