[CDF-26410]  Remove unused constants from ResourceCRUD (#2249)
# Description

Since `v0.4` of Toolkit we have used kind, e.g.,
`my.Transformation.yaml` to identify the resource type of a
configuration file. This means that the constant `filename_pattern` for
`Loaders` is not longer needed. Before `v0.4`, we used this to do regex
pattern matches do guess resource type of configuration files.

Note I make a release just in case here - this should cause no impact to
the end-user, but if it does, a release is good to track down when it
happened.

Note I renamed the `get_loader` to `get_resource_crud` and reimplemented
it with a full test suite. Note I do not test for ambiguous as that is
an impossible state to reach with our current crud classes.,

## Bump

- [x] Patch
- [ ] Skip

## Changelog

### Fixed

- Enforce use of kind to identify resource configuration file. For
example, `my.Transformation.yaml`. This has been an expectations since
`v0.4`.

### Changed

- The `cdf modues update` now requires the current version of Toolkit to
be 0.4 or later.