[CDF-26444] 😃 3D API  (#2250)
# Description

This is preparation for the 3D migration tooling. 

This is reimplementing the 3D integration in CDF. The motivation for
this is the decision we had to leave the PySDK and that the PySDK does
not support the revision response which we need for 3D. Note I only
implemented what I need for the migration.

## Bump

- [ ] Patch
- [x] Skip