[CDF-26411] 📂 Remove list classes (#2262)
# Description

This simplifies the resource CRUD classes by removing the list version
of the request/response classes.

Having these two extra required classes in the CRUDs makes the
implementation more complicated and the benefit is not worth it. By
removing them we reduce the code lines by more than 200 lines.

This will also save us time when we start to move off the PySDK -
instead of implementing 4 classes for each resource, request/response
and list versions, we now only have to implement 2 classes.


## Bump

- [ ] Patch
- [x] Skip