[CDF-26496] ☃Allow tricky metadata. (#2285)
# Description

Realize that the validation for ResourceViewMapping is too strict. There
is no reason to require the mapping to be a valid JsonPath.

The reason is that we are always just constructing a json path not
deconstructing. For example, we can take an asset
```python
{
    "name": "MyAsset",
    "metadata": {"雪ヘ罪約べげド. [10] SNL": "値テスト"},
}
```
and make it into
```python
{
    "name": "MyAsset",
    "metadata.雪ヘ罪約べげド. [10] SNL": "値テスト",
}
```

We never have to go the opposite direction. Thus requiring a json path
is too strict.

## Bump

- [x] Patch
- [ ] Skip

## Changelog

### Improved

- [alpha] When creating a ResourceViewMapping, Toolkit now allows
characters such as `.`, `[`, and `]` in the mapping