[CDF-26481] 🤒 Improve HTTPClient (#2292)
# Description

Improving the `HTTPClient` based on learnings

* Moved `console` to the `__init__` such that it does not have to be
passed in with each request.
* Rewrote all `HTTPMessage` classes to pydantic format. This reduces the
size of those classes in two.
* Added `request_single` to the `HTTPClient` to work with the new
`HTTPMessage` classes.

Too avoid a large PR, I have chosen to build up the new `HTTPMessage`
classes in parallel to the existing code. The plan is as follows
1. (This PR) Introduce new `HTTPMessage` classes for generic request and
responses.
2. Next PR, introduce new `ItemMessage` classes to handle items
requests.
3. Replace all calls to the existing `request` and
`request_with_retries` with the `request_single` and `request_items`
method.
4. Delete the existing `HTTPMessage` and remove the `2` suffix from the
new `HTTPMessages`.


## Bump

- [ ] Patch
- [x] Skip