fix: increase DM query limit for canvas components from default 100 to 1000 (#2334)
# Description

The default limit for NodeResultSetExpression and
EdgeResultSetExpression is quite low at 100 and I encountered a case
where I ended up losing some annotations when downloading and uploading
a Canvas. I believe we could increase the individual limits for each
type of subquery to accommodate for cases where users end up with large
canvases mostly consisting of annotations and so on. Setting it to 1000
for each type of Canvas entity should be safe and does not require any
pagination.

## Bump

- [x] Patch
- [ ] Skip

## Changelog

### Fixed

- Fixed a bug where it was not possible to retrieve all canvas entities
using `cdf download canvas`

Co-authored-by: Magnus Schjølberg <magnus.schjolberg@cognitedata.com>